Check that no key is used for different accessors
The generation of the enum and config class would work just fine with a
duplicate key. However it would not be possible to correctly read an
actual config file with such a duplication. Only one accessor would be
filled in that case.

Therefore we refuse the generation if a key is used multiple times for a
@Coat.Config.