Test that all basic types can be converted
Basic types are:

 - int
 - long
 - double
 - boolean
 - String

All these types are verified as
 - mandatory parameter
 - optional parameter
 - parameter with default value

The code was fixed to support missing conversions.
For example an Optional<Boolean> couldn't be converted. Also hex, binary
and octal ints and longs must explicitly be supported which was not the
case (interestingly for Double not explicit support is necessary as its
valueOf() method already parses these correctly.).