Allow omitting the @Param annotation
If an interface is annotated with @Coat it can be assumed (for the
moment) that all methods should be parameter accessors. Since the “key”
already is optional, the @Param annotation can be optional as well.

In the future we should only automatically consider abstract methods.
This would allow default methods in the annotated interfaces and may
even prepare for the possibility to annotate abstract methods.