Allow reading config values from multiple sources
Config values can now be read from environment variables. Since the
format of environment variables is much more restricted, a relaxed
parsing has to be used.

The generated config object can now be filled with multiple sources. The
order in which to apply them (and therefore which source takes
precedence) is explicitly defined when creating the config object.

To allow for such multiple sources without making the creation of config
objects inconsistent a breaking change had to be made to the generated
classes. Instantiation with constructor is not allowed anymore, but
factory methods have to be used.