Massive refactoring
The processor code was split into several helper classes for easier
maintenane.

A full hierarchical data model is that represents the annotated
interfaces, its accessor methods and its embedded config classes is
built before starting the code generation.

After the full data model is built all assertions are checked and errors
are generated if they are not met. No exception is thrown anymore if the
annotated java files are faulty.

The annotations and accompanying classes were moved to coat-runtime to
avoid a runtime dependency on the processor (and its transitive
dependencies).

The used dependencies in the processor were reduced.

A few more tests were written.