Generate public Builder and valdate on creation
The only publically visible generated class is now the Builder for the
config implementation.

At the same time validation is now done on creating a config object via
the generated Builder. This provides a massive performance boost as all
conversion is done only once on creation time. Calling the accessors
will not trigger any conversion again and is therefor very fast.

This is a breaking change as the ConfigValidationException is now thrown
on calling the #build() method of the Builder and this method only
returns the config interface, not the implementation.

To create the Builder a static #create() method is provided.

Also the global converters and list parsers that can be registered at
runtime are now set via the `Coat` class (which previously was an
annotation).